Update opennft.ui
Since now there is a checkbox "allow for negative feedback" in the UI,  there should also be some cut off for negative values (in nfbCalc.m file). Either you simply use if dispValue< - P.MaxFeedbackVal, or you give the user the option to set the minimum and maximum differently and include a new field for the Minimum Feedback Value in the UI as suggested here.