Update opennft.py
As in the pull request I created for the UI, I suggest to include a field MinFeedbackVal that works in the same way as MaxFeedbackVal and is passed with the struct P to nfbCalc. We have changed our forked version accordingly and it works well (since it follows exactly the some structure as MaxFeedbackVal , the changes should not affect anything else).