Update nfbCalc.m
As in the previous pull requests, I suggest to include a field MinFeedbackVal in the UI and here to cut off the negative values if negative dispValue is allowed. Here, I thus included the case where the feedback value is negative, and smaller than the minimum specified by the user.